# openapi.yml
openapi: 3.0.3
info:
  title: vault-lambda-rs- OpenAPI 3.0
  version: 3.0.3
paths:
...
  /api/v1/record:
    post:
      tags:
        - record
      summary: Create a new record
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordFC'
      responses:
        '201':
          description: Created
      
    get:
      tags:
        - record
      summary: Query records
      parameters:
        - name: 'from'
          in: query
          description: Query records from given date
          schema:
            type: string
          required: true
          example: 2024-04-26
        - name: 'to'
          in: query
          description: Query records up to (but not including) given date
          schema:
            type: string
          required: true
          example: 2024-05-01

      responses:
        '200':
          description: 'Array of found records'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecordQueryResponse'
  /api/v1/record/{sk}:
    delete:
      tags:
        - record
      summary: Delete a Record
      parameters:
        - name: 'sk'
          in: path
          description: Sort key of a record
          schema:
            type: string
          required: true

      responses:
        '204':
          description: 'No content'
...
components:
  schemas:
...
    Record:
      type: object
      properties:
        pk:
          type: string
          example: "Record"
        sk:
          type: string
          example: "2021-08-01T00:00:00Z"
        name:
          type: string
          example: Meal 
        amount:
          type: integer
          example: 3
        unit:
          type: string
          example: kcal
      required:
        - pk
        - sk
        - name
        - amount
    RecordQueryResponse:
      type: object
      properties:
        records:
          type: array
          items:
            $ref: "#/components/schemas/Record"
        from:
          type: string
          example: "2024-04-26"
        to:
          type: string
          example: "2024-05-01"
      required:
        - records
        - from
        - to
    RecordFC:
      type: object
      properties:
        name:
          type: string
          example: Meal
        amount:
          type: integer
          example: 3
        unit:
          type: string
          example: kcal
      required:
        - name
        - amount
...